Fix tests not to rely on returned object order.
LDAP is not guaranteed to return objects in any specific order,
and specifically not in the order they were added in. Therefore,
the tests test_slice() and test_values*() in GroupTestCase fail randomly
depending on the order of results returned by LDAP server.

Instead, sort the results to ensure that we can assert them in specific
order.