Register ExactLookup on lists, along with ListContainsLookup. Closes #105.
Rationale: While __contains sounds like the correct way to find objects
where a ListField contains a value, it does not align with Django's
understanding, e.g.  on a ManyToManyField, where such checks are made using
the ExactLookup.  It also is not natural for people being fluent in LDAP
itself.  I have kept both for backwards compatibility.