-WIP- All tests pass.. but there is a problem with the settings
I managed to fix the tests but even when fixing the settings
save and restore (I missed a copy in the tearDown in previous
commit) restoring the settings raise connection errors.

The way to save settings and restore them is way more complex
than just saving them in setUpClass.

The example of tests without TLS shows why this become mandatory
to use a proper way. Other solution can be to always enable TLS
in tests...