Use _base_manager instead of objects when modifying model
Guard against issues with a non-empty default QuerySet returned by
a Model's Manager instance by calling the Model's _base_manager when
doing lookups by dn. LDAP lookups by dn must be done by setting the dn
as the base of the LDAP query.

The current method for determining if a dn lookup is being requested
relies on inspecting the generated django.db.models.sql.query.Query and
looking for a 'dn' target column in the first where clause statement. If
a Model uses a custom Manager to append a where clause statement to all
QuerySets, this logic is broken. Using _base_manager instead of objects
as the Manager for constructing explicit dn lookups avoids this problem.
This behavior also matches the Django documentation and internals for
related object access, so it is likely to continue to work for the
foreseeable future.

Fixes #196